Lots of updates to the New Order.
main(): Convert to the new Switchboard mechanisms.  This means we
neither need to instantiate the mailing list, nor create a Message
object from stdin.  A side effect of this change is that we can't
efficiently determine whether this message was destined for the -owner
or -admin address (since this script bogusly handles both).  We defer
this decision to the CommandRunner.