Improvements to make sure multiple qrunner masters aren't running at the same time.  Specifically,
main(): Create a master qrunner lock file, with a lifetime of 10 days
(but see below) and exit if that lock cannot be acquired.

master(): Start the lock-refresher subprocess and restart it if it
happens to exit.

start_lock_refresher(): A tiny subprocess that owns the qrunner lock
and periodically (once a day) wakes up to refresh it.