__onefile(): The qrunner framework now sets up the language context for each processed message.  It does this by getting the preferred language of the sender of the message (defaulting to the list's preferred language if the sender isn't a member of the list).
Next, it saves the previous global language context in a local
variable, sets the context to the message's preferred language and
disposes of the message.  Then it restores the original language
context.

It also sets the message's metadata (key: `lang') to contain the
language name for the language context in case any handler modules
need access to it.