Lots of i18n changes.  Specifically,
We have to play module global footsie with _() so that the rejection
notices are marked as translatable, but aren't actually translated
until later.

Administrivia.rejection_notice(), MessageTooBig.rejection_notice():
Convert to interpolating local variables.

hold_for_approval(): Conversion to the new way of i18n'ing.  The
language context at this point is the user's preferred language, so we
craft his notification first before temporarily changing contexts to
the list's preferred (for the admin).  We no longer need to duplicate
the code as was done before, but some things need to be re-translated
in the context of the list's preferred language.

Also, translate the exception str() and rejection_notice for real
here.