Big i18n changes.  Specifically,
send_digests() -> send_i18n_digests(): this does the bulk of the
digest crafting, but it's wrapped in send_digests() which first sets
up the language context for the digest (the list's preferred
language).

send_i18n_digests(): Also make sure that for text/plain parts, the
charset is set correctly to the list's preferred language character
set.

Mark more strings as translatable.