FormatPasswordStuff() -> password_inputs()
membership_options(): We have to urllib.quote() the value attribute on
the hidden input tag, otherwise it is impossible to unsubscribe
addresses with quotes in them (which are valid RFC822 addresses).

change_options(): Be sure to urllib.unquote() the user value field,
undoing the effects of membership_options() and allowing
unsubscription of addresses with quotes in them.