Various fixes for better handling of confirmation messages, specifically:
ParseMailCommands(): Use a mimelib MsgReader idiom to get just the
body of the message, sans headers.  This could be made easier with
Python 2.1's xreadlines module or a better interface on Message
objects.

Substitute one crufty parsing of the Subject: line for another.
Actually now, any command can appear anywhere on the Subject: line,
and confirmation messages are just another form of that (by changing
the Subject: text from "request" to "confirm" -- same as the body
command).

Fixed typos in language support (mlist -> self).

Rewrite the "errors-in-command" string so that the substitution keys
are actually the correct ones!

AddError(): Python 2.0-isms.