Rewritten and simplified.  There's no need to make the logic live inside a class; use the module as if a singleton instance.
The pending database is now stored in data/pending.db.

new(): Stores a new cookie in the pending.db with the key an sha
hexdigest based on the current time, a random number and the content.
The timestamp is now the point in the future at which this entry can
be evicted.

confirm(): Given a cookie, return the data (with the timestamp
stripped off), or None if the cookie is missing from the database.

_load(), _save(): Rewritten and simplified low-level marshal/unmarshal
of the database.  Assumes lock is acquired.

_update(): For use by the bin/update script to convert from
pending_subscription.db to pending.db (the old keys ought to be
compatible).