__init__(): "join" is now a synonym for "subscribe".  "remove" and "leave" are synonyms for "unsubscribe".
ProcessUnsubscribeCmd(): It is now legal to omit both the password and
email address when unsubscribing via the -request address.  If the
password is omitted, a removal confirmation message is generated.
However if the password was given, it must match the user's password
or the removal fails.  If the password matches, the removal is
performed immediately.

Note that there's an ambiguity if exactly one argument is given: is it
a password or an email address?  Use FindUser() to decide giving
preference to email address.