Changes to the new pending.db file format:
- Entry eviction times are stored as a sub-dictionary under a separate
  key called `evictions'.  This leaves the cookie data as whatever was
  given to the new() function.  The keys of the eviction dictionary
  are the cookies and the values are the time at which the entry
  should be removed.

- Another new key `version' for future backwards compatibility <wink>.

- Each entry for the cookie should have a op key as its first element,
  which currently should be either Pending.SUBSCRIPTION or
  Pending.UNSUBSCRIPTION.

new(): Make sure the `evictions' dictionary is present.

confirm(): Don't strip off the last element of the tuple.

_load(): If there's no pending.db file, return a dictionary containing
an empty `evictions' sub-dictionary.

_save(): Check the eviction times in the `evictions' sub-dictionary.
Also, add the schema `version' entry.

_update(): Fix the conversion from pending_subscriptions.db to
pending.db file format so that SUBSCRIPTION entries include a default
language.  Also move the eviction times (which in the old format are
the time the request was made) to the `evictions' sub-dictionary and
add the PENDING_REQUEST_LIFE value.