Support the new confirmation scheme for both subscriptions and removals.  Specifically,
AddMember(): Use the new Pending.new() function interface, which
requires that an appropriate operation key come first.

ProcessConfirmation(): Include more detail on the MMBadConfirmation
exceptions.  Decode the operation key first to see whether we're doing
a SUBSCRIPTION or an UNSUBSCRIPTION confirmation.  The latter is new
and simply calls DeleteMember() if confirmed.

ConfirmUnsubscription(): Similar to subscription confirmation, this
generates a response with the removal confirmation instructions and
url.