GetConfigInfo(): Two new configuration variables have been added:
    - digest_volume_frequency describes how often a new digest volume
      should be started.  Values are Yearly, Monthly, Quarterly,
      Weekly, Daily.  When a new digest volume is started, the volume
      number is incremented and the next_digest_number is reset to 1

    - digest_last_sent_at records when the last digest was sent, used
      to calculate when a new digest volume should be started.

In addition, two new pseudo-variables are added which allow the list
admin to start a new volume with the next outgoing digest, and to send
the current digest right now, if it isn't empty.

send_digest_now(), bump_digest_volume(): New methods to encapsulate
sending a digest and bumping the volume number.