Fixes to handle User-Hits-Stop-Button problems, specifically,
main(): Set up a signal handler to catch SIGTERM, and unlock the
mailing list when this happens.  This has the side effect of aborting
any changes to the MailList object that this web hit may have made.

This is necessary due to semantics of Apache's mod_cgi: when the
browser closes the socket, eventually Apache receives a SIGPIPE (on
output to the closed socket).  This causes Apache to SIGTERM the cgi
process, wait three seconds, then SIGKILL it.  We want to be able to
clean up the locks, so the best we can do is try to unlock the list on
the SIGTERM.  Once we get SIGKILLed, there's nothing we can do.

This change also moves the Save() call into the try: block so that the
finally: block /only/ unlocks the list.  Thus, the list gets unlocked
in most situations.  There are still race conditions where 1) the
config.db file could be corrupted; 2) list locks could still be
unreleased.  Given the semantics of signals in Python, the interaction
of Apache's mod_cgi, and other factors, this is the best we can do,
and it should be better than the old situation.

XXX What do other web servers or cgi execution environments do?