Be more paranoid about abnormal shutdown conditions, specifically:
main(): When the user hits the stop button on their browser during a
long running operation, Apache 1.3/mod_cgi will eventually catch a
SIGPIPE when output is written to the client.  It then turns around
and SIGTERMs the cgi process, waits three seconds, then SIGKILLs the
cgi process.  This patch fixes the stale lock file that can result
under this situation.

Since Python by default doesn't catch SIGTERM, and SIGKILL is
uncatchable, either signal would cause the cgi process to exit without
raising an exception, giving no chance for the script to clean up
after itself.  We now open the MailList object in two phases: first,
open it unlocked, then lock it and install a SIGTERM handler.  The
SIGTERM handler unlocks the list and exits, aborting any changes that
may have taken place.  This is the safest way to ensure that stale
locks won't be left around causing other hits on the list to become
wedged for a long time.