List creation/removal hooks now keep both the plain text file `aliases' and dbhash file `aliases.db' in sync.  Specifically,
_addlist() => addlist()

_rmlist(): Removed.

addlist(): This now takes both a dbhash file and a file pointer to the
plain text file and adds the new list entries into both.  If the plain
text file is empty (seek to end, tell() == 0), then an informative
comment is added to the top.  In the plain text file, `stanza' marker
comments are used around the list alias entries.  A trailing blank
line is always added after the stanza.

create(): This is now just a wrapper for addlist() (since addlist() is
now used by bin/genaliases).

remove(): Does the bulk of what _rmlist() used to do, plus it keeps
the plain text file in sync.  It searches for the appropriate stanza
marker and removes every line between it and the stanza end marker.
(Actually, it does this by copying to a tmp file and shuffling the tmp
file to the real file.)