_dispose(): Add support for mylist-join and mylist-leave aliases.  The mail scripts simply add metadata `tojoin' or `toleave' respectively, which tell CommandRunner which operation to perform.
This is a bit crufty since MailCommandHandler.ParseMailCommands()
still handles the actual job of adding or deleting the member.
CommandRunner hacks the Subject: line of the message to add just the
specific desired command, and it empties the message's payload
(effectively ignoring any additional commands or useless text in the
email message).