InitVars(): Removed, since self._template_dir is no longer necessary and all references to this volatile instance variable have been removed.
SnarfHTMLTemplate(): Also removed, since maketext() does all this in a
more general way.

ParseTags(): Use Utils.maketext() instead of SnarfHTMLTemplate().

InitTemplates(): Because of the new support for expanded template
searches, the only thing that's necessary for a list to support a
language is for it to have a (possibly empty) directory named after
the language code.  Thus, the copying of template files done in this
method are unnecessary.