To main changes: add a -f/--force switch, and scan all the template subdirectories, removing any more-specific templates where identical more-generic templates are available.
Also, more as translatable many more strings.

move_language_templates(): Scan $prefix/en/* for template files.
Search the following locations for identical templates, and when the
md5 checksums match, remove the more-specific versions:

    lists/<list>/<template>
    lists/<list>/<template>.prev
    lists/<list>/en/<template>
    templates/<template>
    templates/<template>.prev

__main__: Added -f/--force option which says to do the update tasks
even if the version numbers indicate a re-install of the same version
or a downgrade to a previous version.  The former case is useful while
developing the alpha versions.  The latter can be pretty dangerous.