A few language related safety issues:
GetConfigInfo(): If list support for a language was removed, but the
list's preferred_language wasn't updated, you could get a ValueError
when setting up the selection widget.  Just use the first entry in the
list.

GetAvailableLanguages(): Ignore the language directory if
LC_DESCRIPTIONS hasn't been updated to include the language.