Fallback module for auto-creation/removal of lists.
When a list is created or removed via the cli (i.e. bin/newlist), this
prints an informative message containing the /etc/aliases entries to
standard out.

When a list is created or removed ttw, this sends an email message to
MAILMAN_OWNER telling him what needs to be done (this isn't ideal but
it's the best we can do).