The add the field (with replacement tags) for suppressing the password reminder for this list.
IMPORTANT: THIS TEXT NEEDS TRANSLATION.

Widen the personal options table to 100% (still not sure about this
one -- it was 80%, but it didn't look good not centered).

Switch the order of the "Don't receive my own mail" buttons.  It was
confusing that this was the only one that was ordered "Yes" then
"No".

ALSO: Un DOS-ified this file; line endings should be Unix-style
only, i.e. \n not \r\n.