New user configurable variables, for control of the (non-templated) web UI colors:
    WEB_BGCOLOR
    WEB_HEADERCOLOR
    WEB_ADMINITEM_COLOR
    WEB_ADMINPW_COLOR

ADMIN_COOKIE_LIFE is removed.  We now only support session cookies.

New authentication role variables:

    UnAuthorized
    AuthUser
    AuthCreator
    AuthListAdmin
    AuthListModerator
    AuthSiteAdmin

With an explanation of what each role is for.  Code will glom together
sequences of authentication roles when they want to do a
web-authentication.  This essentially specifies which
passwords/cookies will be accepted for a particular task.