Document: Set the default background color of the document to mm_cfg.WEB_BGCOLOR.
Document.set_bgcolor(): New method to set the bgcolor attribute.

Document.Format(): Add self.bgcolor to the kws dictionary if it isn't
already there (i.e. allow it to be overridden by the caller).  This
means that we don't need to pass the background color in the Format()
call for most cases.  Cleans up the code and parameterizes the
background color for site customization.