main(): Convert to the new world order for authentication.  We don't use Auth.authentication() anymore because mlist.WebAuthenticate() now does the job nicely, with all the acceptable authentication contexts passed as an argument.
Also, when logging out, ZapCookie() takes an authentication context
(in this case it's AuthListAdmin).

main(), admin_overview(), option_help(), show_variables(),
add_options_table_item(), membership_options(), submit_button(): We
don't need to pass the document background color as an argument
anymore.  Other colors are no longer hardcoded, but taken from mm_cfg.

show_results(): Small change to admindb link text; these are called
moderator requests now.

password_inputs(): Change this section so that it prompts for both
changes to the list owner passwords, and the list moderator
passwords, the latter of which gains access only to the pending
requests page.

change_options(): Because we no longer require the admin password in
order to change the password (because it would have had to have been
supplied in order to even get here), we can simplify the password
change processing.  We also need to add the moderator password change
processing, and this should happen before the list owner password
change processing, because the latter requires that the authorization
cookie be regenerated.