Add support for the list moderator role, and change of address requests.  Specifically,
InitVars(): Add the moderator attribute, which is a list of email
addresses a la the owner attribute.

GetConfigInfo(): Added some details to the `owner' attribute
describing the different list-related roles.  Also added the
`moderator' attribute and copied the description of roles to its
details too.

ChangeMemberAddress(): Added support for change-of-address.  This
method simply pends the request and sends the confirmation email
to the new address (the old one is already verified by way of the web
authentication).

ApprovedChangeMemberAddress(): Effects the actual change-of-address by
adding the member, copying over the user options and the language
settings, and then removing the old address.  Acknowledgements are
always sent.  Handle the global change-of-address by searching for all
the matching lists (must match both host_name and have the old address
of the user as a member), and recursively calling
ApprovedChangeMemberAddress() on those lists -- with the global flag
turned off of course!

ProcessConfirmation(): Add support for CHANGE_OF_ADDRESS, which simply
unpacks the data and calls ApprovedChangeMemberAddress().

ApprovedAddMember(): Fixed a bug in the order of arguments in the call
to ApprovedAddMembers() - plural.

DeleteMember(): Fixed a bug in the i18n-marked subject text, related
to auto-interpolation.

SetPreferredLanguage(): Assert that the language we're setting to is
in the list of available languages (helped catch the
ApprovedAddMember() bug above).  Also, only add
DEFAULT_SERVER_LANGUAGE if it isn't already an item in the list.