A few more updates to get this script working.  Specifically,
main(): Only lock the list during the NumRequestsPending() call.  We
don't actually change the state of the list data, but the ListAdmin
interface requires the list to be locked even to read (yes, this is
bogus).

Also, fix the _() marked string to use the required
auto-interpolation.

pending_requests(): Mark some additional strings for i18n.  Also,
pending subscription requests include the language in the unpacked
data (fix the tuple unpacking).