Classes Document and HeadlessDocument are almost identical, except that the former adds <head><body> tags, while the latter suppresses them.  Document should be used for pages that are created by the scripts out of whole cloth, while HeadlessDocument should be used for any template-based page (which must supply its own head & body tags, although not all of them do, I think).
It turns out to be useful to transform a Document into a
HeadlessDocument, so the trigger to output the head & body tags is the
instance variable "suppress_head" which Document defaults to 0 and
HeadlessDocument defaults to 1.  All the real work is performed in
Document.Format().

Document.get_content_type(): Add the extra newline here, for
convenience.