DEAL_WITH_PERMFAILURES_EVERY: controls how often _doperiodic() will try to deal with deferred permanent failures.
_doperiodic(): It can be a bit expensive to actually deal with
permanent failures every time through this loop, and we don't need
that level of accuracy.  So now we keep a counter and only process
permanent failures every DEAL_WITH_PERMFAILURES_EVERY times.  (Okay,
so this variable defaults to 1, but you get the idea. :)

Also, move the Save() inside the try clause.  I've decided that
finally clauses should only unlock the list, not try to save it.