ConfirmUserPassword(): Obsolete; removed.
AuthContextInfo(): Allow `user' argument to be optional; default to
None.

Authenticate(): Re-install backwards compatibility for pre-MM2.1 list
passwords.  For AuthListAdmin context only, if the sha check fails,
fallback to crypt (if the crypt module could be imported) and md5
checks.  If either of the latter two matches, assume that the list
password is "old style" and auto-upgrade it to the sha password, based
on the matching plain text response.  This way, MM2.0.x list passwords
won't just be broken upon an upgrade.