main(), show_results(): Removed unnecessary global declarations.
main(): If the WebAuthenticate() fails, search for the form key
`adminpw' -- not `admlogin' -- to decide whether to include the error
message or not.  The former is the cgi key for the password field,
while the latter is the key for the submit button, which isn't
included in a Mozilla posted form if the Enter key is hit instead of
clicking on the button.