Lots of changes to make this act more like a traditional daemon and to remove the run-from-cron functionality (yes, I know this script shouldn't be in cron/ anymore, but moving it is a PITA). Specifically,
Removed the SIGCHLD signal handler since this will no longer be run as
a cron job.  Added a PIDFILE global to name the file holding the pid
of the master qrunner process.

sighup_handler(): When the master gets a SIGHUP, send a SIGINT to all
the children so that they'll exit.  This has two benefits: it will
automatically close all open log files, and the auto-restart feature
will mean all the children will see any changes to the source code
(i.e they'll restart and thus import afresh any updated modules).

master(): When the main loop exits, be sure to include the
watchdog_pid in the list of children to SIGINT.

main(): Removed the -b/--background flag.  Write the pid of the
current process to data/qrunner.pid so that it can be exited
externally via: kill -INT `cat data/qrunner.pid`

Convert all syslog() calls to new style.