LogMsg() -> write()
write(): Just call new function write_ex(), passing arguments through.

write_ex(): Use explicit arguments to pass in the args tuple and kws
dictionary.  This is because Python requires **kws in extended calls
to be concrete dictionary objects, not generic mapping objects.
SMTPDirect wants to pass in an instance of (subclassed) UserDict, so
it calls write_ex() directly.

Also, stash the msg argument in local variable origmsg, which is used
if we get an exception during string interpolation.