Added support for an Urgent: header, which should contain a password (either the list administrator's, or the list moderator's).  If the message contains this header with a matching password, the message is sent to the entire list membership immediately, even to those members who receive the message via digests.  Note that the message /is/ still digested too.  Specifically,
RejectUrgentMessage: New class, derived from RejectMessage, which is
raised if the password in the Urgent: header does not match.  Thus,
the original message is returned to the sender immediately, and is not
held or posted.

process(): Check Urgent: for a matching list moderator's or list
administrator's header.  If found, set the recipients list to both the
regular and digest delivery members.  If Urgent: is found, but the
password doesn't match, bounce the message back to the sender.