process(): Be much more careful about bogus source strings for the dictionary interpolation.  Since they're coming from the great unwashed masses, the autoresponse texts could have bogus %()s bits in them, and using a SafeDict doesn't guard against all kinds of exceptions.
So, if any exception is caught during interpolation, log it, and use
the raw autoresponse text in the email.

Also, because we're not using the buggy rfc822.py module anymore, we
don't need to special case the situation when the first line of the
autoresponse text has a colon in it.