process(): Use the MemberAdaptor API.  Also, call do_topic_filters() to further cull the recipients.
do_topic_filters(): For each recipient of the regular delivery
message, if there are no topic hits, check the
ReceiveNonmatchingTopics option to see whether this recipient should
get the message.  If there are topic hits, remove the recipient unless
they've indicated an interest in this topic (or they haven't specified
any interests).