A first shot at configuration for cleaning headers before a message gets gated mail->news.  Specifically,
NNTP_REMOVE_HEADERS: This is a list of headers to remove from the
message before posting to nntp.

NNTP_REWRITE_DUPLICATE_HEADERS: A list of headers that should get
moved if there are more than one in the message.  Contains a tuple of
the original field name and the new field name.