main(): Make the options page a little more bulletproof against membership mining when private rosters are being used.  Also, fix a bug in the chopping up of the url parts.
We now print the "Authentication failed" message in the authentication
clause when the `password' key, not the `login' key is present (the
latter won't be when the email address is given on the url, but the
former will always be present).

When mlist.private_roster is <> 0, this means we do have private
rosters (either to the list membership or to the list admins).  In
that case, set user to None and continue on.  The display will use the
provided email address even if it's not a member, and we'll get a
normal "Authentication failed" message, which doesn't reveal whether
it was the email address or the password that mismatched.

loginpage(): Always set the form action to .../mailman/options
without the email address in the url.  If the user was provided to the
login script, hide it (obscured) in a Hidden input field.  This way,
we essentially clear any unauthorized emails from the url.