process_form(): Lots of cleaning up of the code.  Also, make this page more bulletproof against membership mining for private rosters.
If rosters are private, then we'll just put up a fairly generic
acknowledgement that's the same regardless of what kind of approval is
necessary, or whether the email address is already subscribed or not.
Thus no information about private rosters can leak.

Note also that if a subscription request comes in for an already
subscribed address (using private rosters of course), then we send an
alert to the address to let them know that somebody was trying to
re-subscribe their address or mine the membership.

I also rewrote a bunch of the results strings (sorry translators!)

remote_addr(): Remove this.  We can do it more cleanly inlined.