InitVars(): Remove the initializations for the obsolete list attributes posters, forbidden_posters, moderated, and member_posting_only.  Add initializations for the new sender-centric moderation attributes default_member_moderation, accept_these_nonmembers, hold_these_nonmembers, reject_these_nonmembers, discard_these_nonmembers, forward_auto_discards, generic_nonmember_action.
GetConfigInfo(): Rework this framework so that it doesn't return a
dictionary with keys being the category.  Instead, pass the category
and subcategory in as arguments and let the dispatched-to gui methods
decide whether and what to return in response to these arguments.
This allows the framework to be more robust, and to easily handle
extensions via subcategories (as was done to the Privacy category).

ApprovedAddMember(): Set the newly subscribed member's Moderate flag
to self.default_member_moderation.

ProcessConfirmation(): Argument userdesc_overrides -> context.  In the
SUBSCRIPTION clause, only add context to userdesc if context is an
instance of UserDesc.

Also, add a HELD_MESSAGE clause which handles admin approval/discard
of held messages via email response.  This code is a bit more
elaborate than I'd like it to be, so perhaps it should be factored
out. ;/