Added some internal methods to help with mailmanctl's transfer of lock ownership to its forked children.  You shouldn't normally use these methods, but to make mailmanctl maximally convenient, we need to create a lock in a parent and manage the lock in the child.
_transfer_to(): Called in the parent, this twiddles with lock
internals to set the temp file name to include the pid of the child
(passed in).  It also transfers lock claims to the new temp file,
i.e. the child, and removes the claim made by the parent (the previous
winner).

_take_possession(): Called by the child, this resets the temp file
name to include the pid of the child.  For this, it had the pid of the
parent process prior to the fork.

__del__(): Don't finalize() the lock if we don't own it (ownership is
implied by creation, and relinquished by _transfer_to()).  This way,
when the parent transfers ownership to the child and then exits with
sys.exit(), the act of destroying the lock instance in the parent
won't unlock the lock.

__init__(): Initialize __owned to 1.