A fairly significant rewrite, but now the `restart' command actually does the right thing!  The sub-qrunners are exec'd now from bin/qrunner using some new command line options, so killing and (auto-)restarting them will cleanly reload any changed modules.
Also, the lock is acquired in the foreground so you don't get ugly
error messages if another master qrunner is already running.

Finally, the separate lock-refresher process is gone.  Too hard to
implement correct lock ownership transfer semantics with it (it was
more complicated than the simple pass-thru to the child).  Now, lock
refresh is implemented by a once-a-day alarm signal in the master
qrunner process watcher.

Also implemented the `reopen' command which causes all the log files
to be re-opened (very useful if you're rotating log files!).