A major rewrite to 1) better support the new mailmanctl script and 2) to hopefully lay the groundwork for a MM2.0.x compatible cron-invoked qrunner process.
This qrunner responds to SIGINT, SIGTERM, and SIGHUP as described in
the mailmanctl docstring.  It also added a more robust -r/--runner
flag to support mailmanctl invoking multiple qrunners per queue via
slicing.

The default now is to run continuously, unless -o/--once is given, and
if multiple -r/--runner flags are given, they will be run in
roundrobin format.  This should be all we need to implement a
backwards compatible cron/qrunner script.