Much revision based on good feedback from mailman-developers. Specifically,
process(): Instead of a blanket discard of text/html parts, what we do
depends on the value for ARCHIVE_HTML_SANITIZER.  Also, the "scrubbed"
message now includes the value of get_filename() if available.

save_attachment(): Several refinements including: all of a message's
attachments are stored in a subdirectory off
archives/private/mylist/attachments.  This subdir is calculated based
on the SHA1 hash of the Message-ID:

We store the attachment in a file based on the name of the
get_filename() value, sanitize for nasty characters, absolute paths,
etc.  The filename is uniquified within the message's attachments
subdir.

This function also knows about text/html and performs filtering when
ARCHIVE_HTML_SANITIZER is a program string.  Note that currently, we
assume that this program will filter html into plain text, so we
change the attachment's suffix to .txt accordingly.

We don't need the attachments.pck file anymore.