process(): Teach the scrubber about message/rfc822 types, which are not multipart, but don't contain a string payload.  They have a single Message instance payload so they need to be scrubbed.
I don't know whether what we do is the best thing, but we strip out
the contained message, store it as an attachment (with very little
processing), and include a link/info block of text with the subject,
sender, date, size, and url.

save_attachment(): When calculating an extension for an unknown type,
default to .txt for message/rfc822 types and .bin for everything else.

Also, if we're saving a message/rfc822 type as an attachment, all we
do is take the raw text of the message, cgi.escape() it for safety,
and store it in the attachment file.  We could probably do better.

Finally, adjust the baseurl for private archives so we don't get a
double slash.