Updates to work better with virgin invocation.
_zapfile(), _zapdb(): Only truncate the file if it already exists (so
we don't accidently create it before we're ready to).

addlist(): Make this robust for when mlist == None, as it will when
there are no lists yet, but we want the preamble and minimal aliases
and aliases.db files to get created.  Moved the YP_* key settings up
before the predicate shortcircuit.

do_create(): Do the hashopen() before the built-in open() so it gets
created properly by bsddb when it doesn't yet exist.

create(): Check for mlist == None before digging out more attributes.

Also, rewrote the big comment at the top concerning locking.