Mailman re-uses connections to the SMTP server when delivering multiple chunks.  Marc Merlin points out that multi-transaction sessions aren't handled correctly; this fixes the problem. Specifically,
process(): In the non-threaded delivery stanza, be sure to quit the
connection when finished delivering instead of just closing it.  This
ensures an SMTP QUIT command is sent.

threaded_deliver(): Same for thread-shared connections; quit instead
of just close.

deliver(): Get rid of the try/finally wrapper around the
conn.sendmail() call.  We definitely don't want to QUIT the session
here.